LibLouis 3: Add pattern.c to list of files to be compiled.
One of the enhancements in LibLouis 3 is centralized pattern matching and translation routines. Among other things, it allows emphasais to be handled better.
Thus add pattern.c to list of files to be compiled as part of compiling NVDA from source, as without this, MS Linker (link.exe) presents unresolved symbol error.