Add-ons manager: refactored how needs restart flag is set by checking for cardinality of sets and comparing sets. re #3090.
As noted by Michael Curran: it is better to make needsRestart a property. However given that it is consulted only when closing, it is better to refactor by:
* Check if any add-ons have been installed or removed ( tehnically, pending).
* Check if disabled add-ons and pending disable set are the same (for the most part, it is).
This also fixes an issue where if restart dialog is dismissed when closing add-ons manager, and add-ons manager is opened again, restart prompt would not be displayed.