LibLouis 3 transitoin: provide a way to migrate old table names to new filenames. re #6140.
Given that people would be surprised if some tables would revert back to defaults, it would be handy to provide a short migration routine to map old table filenames to new ones. Thus, a tempoary routine and a map is introduced to handle it (to be removed at a future date).