ux improvements
When adding a dictionary entry with invalid regex, the error message now
states that this is a regex error and qoutes the error. It is
unfortunate, but the regex error is difficult to translate.

If a user does manage to introduce an error into a dictionary file, then
the nvda will import all valid lines in the file, reporting warnings to
the log for lines in the dictionary that are invalid.