When disabling and then immediately re-enabling an add-on or vice versa, the add-on status now correctly reverts to what it was previously.
When an add-on is in the pending disable set and the user enables it, the user is just undoing the disable; i.e. there is nothing to do. Therefore, just remove it from the pending disable set; don't also add it to the pending enable set. Make a similar change for undoing an enable.