Work around some bugs in Mozilla's IAccessibleText implementation.
When moving the caret forward by word while editing in Mozilla applications, punctuation at the start of a word is no longer ignored. Re #3337.
When reading by line while editing in Mozilla applications, if a list begins on the next line, the list is no longer reported as part of the previous line.