Fix for #6350
Split the welcome part and the details description, in order to provide
different formatting. The 'Welcome to NVDA!' part is larger and bold. I
have tested that both bits of text are reported when the dialog is
opened.

Made the group box fill the width of the dialog, I think this is a more
usual layout.