Fix failure to detect focus change when opening a book from the Kindle library.
IAccessibleHandler.OrderedWinEventLimiter: Don't drop focus events if we have a foreground event with the same parameters. Otherwise, we might drop a valid focus event even though the foreground event is invalid. With this change, pumpAll rejects the invalid foreground event and falls back to the valid focus event.
This will cause us to end up with more focus events in the limiter, so bump maxFocusItems from 3 to 4 to compensate.