Experiment to report the column number (#6559)
Now the number of text columns and the column number are reported in MS word.

This is done by getting the position relative to the start of the document (horizontally), and comparing it to the start position of the text column.
Since word does not have an interface to get the start position of the text column it is calculated. This is done by adding gutter, and margin, then the width of the column and space between to get each successive column.