What's New: Remove mention of support for the upcoming version of the Kindle app.
The version of Kindle supported by this code will not be released. Instead, a later version will be released with more features, but this also requires newer NVDA code which will not be included as part of this release.
This older NVDA code won't cause any major problems, but it won't work as expected. Therefore, we shouldn't mention this in the What's New, but there's no need to remove the code at this late stage.
Re #6247.