Welcome dialog text cut off
Fixed an issue that caused some text in the Welcome dialog to be cut off
when the system language was set to Bulgarian.

A different font was being used (due to `wx.DECORATIVE`) when the system
locale was set to Bulgarian. This is now `wx.NORMAL` which looks fine in
both English and Bulgarian. The labels are added with `wx.EXPAND` so that
they fill their container in case there are some other languages where
this is an issue.

Fix for #6599