Introduce config/profile schema upgrade framework
When a schema is updated, config files (base config and profiles)
can be upgraded to meet the new schema. The `schemaVersion` (in
configSpec.py) should be incremented, and a new function should be added
to the `profileUpgradeSteps.py` file to make the changes required to
config files before they are loaded.

After all upgrade steps are made the file is validated, and any errors
will be in the log.