appModules/explorer: suppress announcement of 'pane' when minimizing windows or moving to desktop. re #6671
Somehow, when pressing Windows=M or Windows+D, NVDA says, 'pane'. It turns out there is a window with the window class name of 'WorkerW' that presents a focus event. Thus suppress this, which means NVDA will no longer announce 'pane' when minimizing Windows. This bug affects various versions of Windows (tested with Windows 7 and Windows 10).