Report sections and columns in MS word - PR #6498
Fixes #5946.

This behaviour is tied to the report page number setting (available in document formatting settings dialog).
When reading through a document:
- report when the section number changes.
- the type of break ( 0x0c char ) is now announced, this may be a continuous section break, a new column section break, new page section break, even pages section break, odd pages section break.
- column breaks ( 0x0e char) are also announced.
- report the number of text columns and the current column.

Reporting the text columns has some fragility. It is done by getting the position relative to the start of the document (horizontally), and comparing it to the start position of the text column.
Since word does not have an interface to get the start position of the text column it is calculated. This is done by adding gutter, and margin, then the width of the column and space between to get each successive column.