Quick nav:
Report sound with NVDA's disable quick nav feature.
Report "Browse mode, quick nav off", "browse mode", "Focus mode" when the mode changes.
For example, suppose the user is in focus mode on twitter, has quick nav disabled on gmail, and is in browse mode on nvaccess.org and xkdc.com
1. User switches from Twitter to Gmail. NVDA says "Browse mode, quick nav off" or plays the disabled quicknav sound.
2. User switches from GMAIL to the NV Access website. NVDA says "Browse mode". or bonks as usual.
3. User switches to Twitter again: NVDA says "Focus Mode" or focus mode bonks.
4. User switches to xkdc: NVDA says "Browse Mode" or browse mode bonks.
5. User switches from xkdc to gmail. NVDA says "Browse mode: Quick nav disabled" or bonks the quick nav off bonk.

Note that a sound is needed for the quick nav off case still.