Add-on handler: add a new function to check for add-on updates, add-ons manager progress dialog changes. re #3208.
Add-on handler: add a new function (checkForAddonUpdates) that'll allow NvDA to check for add-on updates. This is to be used from add-ons manager (manual) or when NVDA starts (automatic). The goal of this function is to gather add-on listings, connect online to retrieve add-on update information, and return this info in a format that a new add-on update result dialog will use to display and download add-on bundles. The add-on update check should not run if NVDA itself is being updated (NvDA Core takes precedence).
In add-ons manager, the progress bar dialog used for add-on installation is now an attribute of the dialog itself, as it is now used for update status message as well. Also, update check button briefly displays progress dialog and returns to add-ons manager. Eventually a new add-on update result dialog will be added that'll display add-on update info returned by addonHandler.checkForAddonUpdates and download new bundles.