Add-ons Manager: provide a function to update add-ons with no prmopts. re #3208.
Similar to add-on instalaltion routines but with a difference: no prompts will be shown when updating add-ons unless serious errors occur. The new updateAddons function takes a map of new add-ons to be installed, which is a map with add-on identifiers (names) as keys and info such as path as subkeys. This function will be called when user tells NVDA to update add-ons from update results dialog.