Add-on handler: introduce a blueprint of add-on update procedure that involves sending and receiving data in JSON format. re #3208.
Javascript Object Notation (JSON) is a handy way to serialize and deserialize data across various systems, including sending data over a network. It can be used to send an receive various data, such as nested dictionaries, lists and so on.
In case of add-on update checker, JSON wil be employed to store add-on update information from a local NVDA installation, as well as to parse what the update server says about a possible update. A blueprint (pseudo-code) has been provided to spell out what the add-on update checker should do to accomplish its task.