Add-on update downloader: initial implementation, add-on download urls will be fetched from info url value. re #3208.
An add-on downloader, based on updateCheck's downloader has been implemented. The only differences are title and content of the progress dialogs, destination path name and extension, and what happens once download completes (installs add-ons).
In terms of update URL's, the same approach as in NVDA Core: url list will be stored in the update info dictionary.
The only showstopper is the problem where progress dialogs for add-on downloads won't go away if multiple add-ons are being downloaded and if at least one of them has an installation UI. Once this is fixed, the manual add-on update check is considered complete (the remaining item for the client is automatic update checks, and the rest will be server's job).