Update add-ons client: added workarounds for add-ons manager focus problem, update results diallg will now handle the case of no add-on updates. re #3208.
Somehow, either because of lack of proficiency in GUI programming (from me) or what not, when add-on update results are displayed, add-ons manager is not focused once that window is closed. This is more so after multiple add-ons are updated at the same time.
Also, update results dialog will present a message (not a list of add-ons) when no add-on updates are available, as well as laying the foundation for automatic update checks for add-ons via 'auto' variable.