Add-on updates: use generators in order to update add-ons one at a time. re #3208.
Noted by Jamie Teh (NV Access) and Leonard from Netherlands: generators are great for doing one thing at a time. For this task, an updateAddonsGenerator will be used to update (download and install) add-ons one at a time. The generator will receive a list of add-ons to be updated, the add-on info for the to be updated add-on will be gather via list.pop, then add-on downloader and instlaler will be told to update add-ons via two kwargs: auto (automatic updates) and and list of add-ons to be updated.

This means:
* While add-ons are being updated, add-ons manager will be destroyed.
* Add-ons will be updated one at a time.
* Success/error methods in add-on downloader will call the generator defined above.