Fix intermittent null context handle exceptions when navigating between pages in Firefox with braille enabled.
If a buffer notifies of an update just before it is terminated, it's possible that the update handler method gets queued but the buffer gets unloaded before the handler runs. In that case, the vbuf handle will be null. The update handler method now checks for this and returns early if it occurs.