Extraxt empty checking logic to NVDAObject
In focus mode we now check if the content of the field is empty via the
NVDAObject _get__isTextEmpty method.

In browse mode, extracted and shared code to get the placeholder
attribute if the content of the field is empty. This is used by IE,
Firefox, and Chrome.

Fixed up some braille and speech cases.