Fix issue in edge focus mode.
Moving by character was not limited to the content of the field, this
meant that the field boundary could be overrun.

The algorithm used could also cause a problem when the use entered
something like "\n Hello there" which would be considered empty.

Instead we use the object to create the text info, this should know the
correct range for the textinfo. Doing this I have confirmed that in Edge
in focus mode, using the test cases outlined on PR #7055 that an empty
field only ever has the value of '\n'.