UIA/Search field: do not announce search field entry when searchui (start menu) is closing. re #7370.
due to use of announceNewlineText, when start menu closes and when Enter is pressed, search field value will be announced, which sometimes leads to side effects such as core freeze. Address this through initOverClass in UIA.SearchField that turns this flag off if and only if we're in start menu in Windows 10.