UIA/ToastWin10/build 15063 and later: before announcing toasts, consult last toast message to make sure it is different, caused by possible problem in uIA Core in Creators Update and later. re #7128.
According to mick Curran (NV Access), UIA Core may have bugs when it comes to firing multiple toast events after using UWP's for a while, so detect this by consulting last toast message if the build is 15063 or later. This does not affect older releases such as Anniversary Update (build 14393).