recogUi: Queue the focus event to the main thread, rather than running it on the current thread.
UWP OCR calls the result callback in a background thread, but NVDA events must never be handled in a background thread.
This was causing MSHTML vbufs to die due to COM queries failing because they were on the wrong thread.
Added a debugWarning in IAccessibleRole which would have made this easier to catch.
Fixes #7399.