Apply change to all scintilla editors
Since there are several other editors that are suffering from the same
bug (Notepad2, PureBasicIDE). Apply the change to all scintilla editors.
This may mean that in some Scintilla based editors the initial problem
of braille not updating as characters are typed with left to right
languages in a right aligned document (or right to left languages in a
left aligned document) will persist. So far however, we are unaware of
any scintilla based editors that suffer from this problem.