Add generic framework to enable extensibility at specific points in the code by way of actions and filters.
This allows interested parties to register to be notified when some action occurs or to modify a specific kind of data.

There are two types of extension points:

1. An Action might be used to notify interested parties that the configuration profile has been switched.
2. A Filter might be used to allow interested parties to modify spoken messages before they are passed to the synthesizer.

Handlers for extension points are always called in the order they were registered so that the order is determinate, which should make behaviour more predictable and easier to reproduce.

Handlers can be called with keyword arguments.
If a handler doesn't support a particular keyword argument, it will still be called with the keyword arguments that it does support.
This means that additional keyword arguments can be added to actions and filters in future without breaking existing handlers.