OneCore Voices: Fix failure to speak some messages when several messages are spoken in quick succession. (PR #7494, issue #7473)
This occurs, for example, when toggling a check box with the space bar with speak typed characters turned on. In this case, often "space" would be spoken, but not "checked"/"not checked".

#7463 (cc0aa5d1) modified the oneCore driver to call  player.idle  when there was no more speech in the queue in order to fix audio ducking. Unfortunately, i neglected to realise that because  player.idle  can take some time, there might be more speech queued while it was running.
Now, we still check the queue and only call  player.idle  if it is empty. However, we then process the queue after that so that if new utterances arrived in the queue during the  idle  call, we still process them.