OneCore Voices: Fix lag when reporting characters.
When we speak characters, we include indexes, which get translated to SSML marks. Unfortunately, calling WaveOutOpen blocks for ~100 ms if called from the callback when the SSML includes marks. We're not quite sure why.
To work around this, open the device before queuing.
Move the call to player.idle to _processQueue so that it gets called (thus closing the audio device) even if the UWP code encounters an exception.