Fix wx warning and GUI stutter when loading
There was a warning to do with the span used. The span should be
(rowspan, colspan) where (1,1) is a single cell.

There was a stutter when changing to certain (more complicated) settings
panels. For instance the Braille settings or the document formatting
settings. Calling `Freeze` and then `Thaw` addresses this.