Fix the unit tests.
The previous commit broke the unit tests because getCurrentLanguage
needed a synthesizer to be set but it wasn't in the test environment.
Fixed the getCurrentLanguage method to not require a synthesizer and
fall back to using NVDA's language. when a synth is not available, just
as if it didn't declare what language it was using.