Prevent freezes on every focus change if Firefox stops responding.
Previously, if Firefox (or Chrome or any other app based on these) stopped responding, NVDA would often freeze every time the focused changed, even if the browser was in the background.
This occurred because NVDA queried the document on every focus change to see if it was alive, but that query blocked in the case of an unresponsive process.

Now, we don't query if the process is in the background.
We just assume the document is alive in this case.
The chances of a document dying while the browser is in the backgroud are low anyway.
However, if this did occur, the buffer will still be killed once the browser comes to the foreground or gets exited.