Touch handler and GUI: check if touch is supported, as touch handler might not be ready when initializing GUI subsystem. re #7309.
Touch handler is initialized AFTER GUI, so handler might not be ready at all. To avoid this, a new touchHandler.touchSupported function will be used to check if touch interaction is supported. For now, it is used by GUI, but add-ons and other modules are more than welcome to use it."