Add debug logging for the time taken to load a virtual buffer.
When testing browser and virtual buffer backend performance, it's extremely useful to be able to measure how long it takes to load a buffer.
Currently, it's possible to achieve this by looking at the IO timestamps and doing manual math, but this is tedious at best.
This logs a simple debug message indicating the time taken and number of characters in the buffer.