Stop alt+a from silently applying settings
When using the wx function "CreateButtonSizer", enter triggers ok,
esc triggers cancel, and alt+A triggers Apply.

We have some conflicts with alt+A, used as accelerators for options.
When there are conflicts for checkbox items the focus is only shifted to
the item, the item is not activated. When the item is a button, the
button is activated. The only feedback a user is given for the apply
button is that focus shifts to the categories list.

This commit stops the settings dialogs from using the
"CreateButtonSizer" method, and instead we create the buttons ourself.
We add a binding for the "enter" key so that enter will still trigger
the OK button. We also add a binding for "control + S" to trigger the
apply button.

To resolve the annoying windows chime when pressing enter on the
synthesiser name, or braille display name, the "...Change" button is now
triggered.

This also means that OK, Cancel, Apply can be translated for all our
users.