UIA constants: remove constants added in current interface that ships with NVDA. Re #8416.
In the old days, when UIA handler supported Windows 7 and 8.x (IUIA1 through IUIA3), there were constants that were not part of the interface file that came with NVDA. now that iUIA5 is supported (and in the future, IUIA6), it makes no sense to continue to include constants that are now part of the interface itself. Thus constants included in latest UIA interface are no longer part of NVDA's own UIA handler module (not all are gone, however).