implemented reviewer comments
if statements splitted in 2 lines and translator lines defined.
At the reviewer and his first comment, that was for another issue regarding the outlook.py and came in here by mistake.
This is the TotalCmd.py only and has nothing to do with the outlook.py, just a wrong commit from my side, sorry.