UIA: assume Windows 7 is in use and remove constants that are now part of current interface (#8422)
* UIA handler: just assume UIA is available, but keep the UIA enabled key. Re #8415.

UIA support in NVDA assumes Windows 7 or later. Because NVDA supports Windows 7 (SP1) and later, just assume UIA is available. However, for cases where UIA should be turned off, turn this on if UIA enabled key (from config) is on.

* Config: remove minWinVersion key for UIA. Re #8415.

As NVDA supports Windows 7 (SP1) and later, there's no need to check if min Windows version is 7 anymore, thus the coresponding key in Config manager/UIA section is no longer applicable.

* UIA constants: remove constants added in current interface that ships with NVDA. Re #8416.

In the old days, when UIA handler supported Windows 7 and 8.x (IUIA1 through IUIA3), there were constants that were not part of the interface file that came with NVDA. now that iUIA5 is supported (and in the future, IUIA6), it makes no sense to continue to include constants that are now part of the interface itself. Thus constants included in latest UIA interface are no longer part of NVDA's own UIA handler module (not all are gone, however).

* Update what's new