EdgeNode/live region and system alert events: address review actions.
Requested by Mick Curran (NV Access): what if authors provide cusotm ARIA labels for the element in question? As of now, the name property does not take this into account. Instead, handlers for UIA system laert and live region change events will be defined with room for expansion when the need arises.