Add-on update checks: add ability to select which add-ons should be updated when update dialog is shown. Re #3208.
Add checkboxes (list control) to add-on updates dialog so users can select which add-ons should be updated. By default, checkboxes are not checked, but one can check add-ons one wishes to update.
Also, whenever checkboxes are checked, check if update button should be enabled or not based on if at least one add-on is checked for update installation (via Python's any() function).