Allow pass through in ARIA menus, fixes #3215. (#8835)
* Allow pass through in ARIA menus, fixes #3215.

In addition to the existing elements like TableCell, column and row headers, allow menu items, menu item radio button and menu item checkbox roles to pass keys such Escape through as well. This should make the interaction in mega menus and other menu types on the web much more intuitive and eliminate the need to press Escape twice to back up a menu level or close a menu alltogether.

* Fix bug for real this time.

The previous commit was total bogus and in the wrong position. The script for the Escape key needs to handle these exceptions all by itself so it no longer almost unconditionally invokes browse mode. Tested and it works. :)

* Make the roles a frozen set

* Address review comments

* Update what's new.