Fix handling of table cells without a containing table in browse mode.
In the new Gmail, message bodies are contained within a "gridcell" which does not have a containing grid.
The gridcell's name is the entire body of the message.
When using quick navigation to move to anything inside these "cells", NVDA reports information about the cell, including its name, after the element to which you just navigated.
This results in a huge amount of spurious (and confusing) verbosity in Gmail when using quick navigation.
To fix this, cells outside of tables are now treated as layout.