Gecko vbuf: Render the selected item in list boxes and trees.
If there is no selected item, render the first child.
This allows the user to see the selected item without having to switch to focus mode.
It also allows focus to move into list boxes and trees where the container itself isn't focusable but the selected item is, as permitted by the ARIA spec.

This required changing speech.getControlFieldSpeech so that the content of lists and trees is spoken after the control field info for REASON_FOCUS.
Otherwise, moving to a list box or tree with quick navigation would speak the selected item first.
List item also had to be added to the roles which always enable pass through, as this change means that a list item can now be under the cursor when enter is pressed. (Read only list items are still excluded by an earlier check.)

This is currently disabled for HTML select size>1 controls in Chrome.
These list items get the focusable state but setting focus programmatically does nothing.
Therefore, we don't want to render these in Chrome because a user wouldn't be able to focus these list boxes in browse mode if we did.